pyproject.toml
[project]
name = "axon"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Zak Johnson", email = "me@zakj.net" }]
dependencies = [
    "textual>=0.72.0",
    "markdown-it-py>=3.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.1",
    "textual-dev>=1.5.1",
    "pytest>=8.2.2",
    "pytest-mock>=3.14.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/axon"]
